remove getSoftwareFormAutoFillDataFromExternalAndOtherSources from redux
