+++ api/src/core/index.ts
@@ -1,2 +1,2 @@
 export { bootstrapCore } from "./bootstrap";
-export type { Core, Context } from "./bootstrap";
+export type { Context } from "./bootstrap";
