--- api/src/core/usecases/suggestionAndAutoFill/index.ts
@@ -1,3 +1,0 @@
-export * from "./state";
-export * from "./selectors";
-export * from "./thunks";
