--- api/src/core/usecases/suggestionAndAutoFill/selectors.ts
@@ -1,5 +1,0 @@
-// import type { State as RootState } from "../../bootstrap";
-// import { createSelector } from "redux-clean-architecture";
-// import { exclude } from "tsafe/exclude";
-
-export const selectors = undefined;
