--- api/src/core/usecases/suggestionAndAutoFill/state.ts
@@ -1,3 +1,0 @@
-export const name = "suggestionAndAutoFill";
-
-export const reducer = null;
